Remove bertscore calculation and other results fetching
