+++ services/QuillLMS/engines/evidence/spec/services/evidence/research/gen_ai/confusion_matrix_builder_spec.rb
@@ -5,37 +5,37 @@ require 'rails_helper'
 module Evidence
   module Research
     module GenAI
-      RSpec.describe OptimalAndSuboptimalResultsBuilder, type: :service do
+      RSpec.describe ConfusionMatrixBuilder, type: :service do
         subject { described_class.run(llm_examples) }
 
         context 'when there are no llm feedbacks' do
           let(:llm_examples) { [] }
 
-          it { is_expected.to eq(confusion_matrix: [[0, 0], [0, 0]], accuracy: nil) }
+          it { is_expected.to eq [[0, 0], [0, 0]] }
         end
 
         context 'with test_optimal, llm_optimal' do
           let(:llm_examples) { [double(test_optimal?: true, optimal?: true)] }
 
-          it { is_expected.to eq(confusion_matrix: [[1, 0], [0, 0]], accuracy: 1) }
+          it { is_expected.to eq [[1, 0], [0, 0]] }
         end
 
         context 'with test_suboptimal, llm_optimal only' do
           let(:llm_examples) { [double(test_optimal?: true, optimal?: false)] }
 
-          it { is_expected.to eq(confusion_matrix: [[0, 1], [0, 0]], accuracy: 0) }
+          it { is_expected.to eq [[0, 1], [0, 0]] }
         end
 
         context 'with test_optimal, llm_suboptimal only' do
           let(:llm_examples) { [double(test_optimal?: false, optimal?: true)] }
 
-          it { is_expected.to eq(confusion_matrix: [[0, 0], [1, 0]], accuracy: 0) }
+          it { is_expected.to eq [[0, 0], [1, 0]] }
         end
 
         context 'with test_suboptimal, llm_suboptimal' do
           let(:llm_examples) { [double(test_optimal?: false, optimal?: false)] }
 
-          it { is_expected.to eq(confusion_matrix: [[0, 0], [0, 1]], accuracy: 1) }
+          it { is_expected.to eq [[0, 0], [0, 1]] }
         end
 
         context 'with mixed feedbacks' do
@@ -48,7 +48,7 @@ module Evidence
             ]
           end
 
-          it { is_expected.to eq(confusion_matrix: [[1, 1], [1, 1]], accuracy: 0.5) }
+          it { is_expected.to eq [[1, 1], [1, 1]] }
         end
       end
     end
